+++ Packages/com.espale.utilities/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2026-02-2
+
+### Added
+
+- Added a new component called `AnimationCurveRenderer` for drawing `AnimationCurve`s with a `LinerRenderer`.
+- Added 2 new constants, `DT_30FPS` and `DT_60FPS`, with the values of delta time of the coresponding frame rate.
+
+### Fixed
+
+- Fixed `BetterDebug.DrawLine` and `BetterDebug.DrawRay` ignoring the duration parameter.
+
 ## [1.0.8] - 2025-09-18
 
 ### Fixed